Update FileDataProvider to use the new model.Unmarshaler, avoid InternalRep usage
